add ability to customize icons in autocomplete menus, fix broken datagrid bg-color alterations after sorting, add some more icons, updated user config to v4, fix unhandled exceptions, add filecopy-task, fix entry validation logic in datagrid

+++ MyCBZ/Result/TaskResult.cs
@@ -12,9 +12,9 @@ namespace Win_CBZ.Data
 
         public Dictionary<string, object> Payload { get; set; } = new Dictionary<string, object>(); 
 
-        public int Total { get; set; } 
+        public long Total { get; set; } 
 
-        public int Completed { get; set; }
+        public long Completed { get; set; }
 
         public TaskResult() { }
